remove test case to integration tests
