+++ libs/community/tests/integration_tests/embeddings/test_qianfan_endpoint.py
@@ -1,4 +1,7 @@
 """Test Baidu Qianfan Embedding Endpoint."""
+from typing import cast
+
+from langchain_core.pydantic_v1 import SecretStr
 
 from langchain_community.embeddings.baidu_qianfan_endpoint import (
     QianfanEmbeddingsEndpoint,
@@ -38,3 +41,17 @@ def test_rate_limit() -> None:
     assert len(output) == 2
     assert len(output[0]) == 384
     assert len(output[1]) == 384
+
+
+def test_initialization_with_alias() -> None:
+    """Test qianfan embedding model initialization with alias."""
+    api_key = "your-api-key"
+    secret_key = "your-secret-key"
+
+    embeddings = QianfanEmbeddingsEndpoint(  # type: ignore[arg-type, call-arg]
+        api_key=api_key,
+        secret_key=secret_key,  # type: ignore[arg-type]
+    )
+
+    assert cast(SecretStr, embeddings.qianfan_ak).get_secret_value() == api_key
+    assert cast(SecretStr, embeddings.qianfan_sk).get_secret_value() == secret_key
